refactoring Study point service
